Added RepeatNode and some tests for it.
As a concequence we also change the grammar a bit so the repeat node has a primaryExpression
as its LoopExpression to make the repeat node grammar similar to the if node.
We also changed the test.alang file, so we had to change the test of location to fit the new file.

Co-authored-by: Mai-Britt Lerche Laursen <malaur18@student.aau.dk>