Added more implementation to IfNode and expressionNode
This commit ended with a failing test for ExpressionNOdes because we couldn't figure out the expected behaviour of ExpressionNode in the AST.
With the Syntax Tree Having only one kind of expression complicates things.. Maybe we need to change the grammar for it.

Note: Also updated the grammar for IfStmt and the test file to match the newly discussed changes.

Co-authored-by: Mai-Britt Lerche Laursen <malaur18@student.aau.dk>
Co-authored-by: Kasper Stenholt Winkler <kwink18@student.aau.dk>