Added precedence into the grammar + added the visit function for it in AstBuilderVisitor.cs.
All tests passes except the last type-checking of assignments. I have added so that the AST have the same nodes as before, but now the ExpressionNodes has precedences in the structure, this makes it easier to type-check them.

NOTE: The code for the AstBuilderVisitor.cs now has a lot of repeating code which can easily be refactored. We have test for some expression cases, but before this is refactored, we should add more tests.

/Smed